Add optional bounds/contract checking
Default behaviour is to mimic assert(), i.e. terminate on violation
unless NDEBUG is defined, in which case the checks vanish.

If TCB_SPAN_TERMINATE_ON_CONTRACT_VIOLATION is defined, it will check
even in the presence of NDEBUG. If TCB_SPAN_THROW_ON_CONTRACT_VIOLATION
is defined, we throw a `logic_error`-derived exception (regardless of
NDEBUG). Lastly, if TCB_SPAN_NO_CONTRACT_CHECKING is defined then
checks will not be used, even when NDEBUG is not set.

Under C++11, things are a bit different, because we can’t have both
constexpr constructors/members and contract checks at the same time. As
of this commit the choice is in favour of constexpr (so by default
there are no contract checks under C++11), but this can of course be
overridden by one of the above macros.