Fix hard error with libstdc++ vector<bool>
For some reason, libstdc++'s vector<bool> specialisation has a data() member, even though it's not supposed to. Because this returns void, it tripped up our constructor contraints and caused a hard error by attempting to form an array of void.

The fix is to proceed with the convertibility check only if data() does not return cv void.

Fixes #24